Use public IP as p2p host in Pantheon config
Configure Pantheon with its public IP address as p2p host instead of 0.0.0.0.
This is because contract permissioning rules match the whole destination enode
including the IP address, and we're whitelisting enodes with their public IP
address.

From an Ansible perspective, this makes the file a template as it's now using
the node_ip variable.